requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.7.0
idna==3.10
iniconfig==2.0.0
lxml==6.0.1
magic-filter==1.0.12
multidict==6.6.4
mysql-connector-python==9.5.0
packaging==23.1
pluggy==1.2.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.2.3
yarl==1.20.1
